refactor: move tabs core CSS files to styles folder
