+++ packages/tabs/src/styles/vaadin-tab-styles.d.ts
