+++ packages/tabs/src/styles/vaadin-tab-styles.js
