+++ packages/tabs/src/styles/vaadin-tabs-styles.d.ts
