+++ packages/tabs/src/styles/vaadin-tabs-styles.js
